backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==5.1.0
mmh3==5.2.1
fastembed==0.8.0
numpy>=1.26.0
python-dotenv==1.0.1
aiohttp==3.11.11
httpx==0.28.1
aiofiles==24.1.0


